tweaking modules for first merge
