+++ dash/SQLConnection.py
@@ -4,7 +4,7 @@ import pandas as pd
 class SQLConnection():
 	"""Create a SQLite connection class
 	"""
-	def __init__(self, database_file: str) -> None:
+	def __init__(self, database_file: str) -> pd.DataFrame | None:
 		self.conn = sqlite3.connect(f'{database_file}')
 		self.curs = self.conn.cursor()
 	
@@ -21,24 +21,6 @@ class SQLConnection():
 		try:
 			res = pd.read_sql_query(query, self.conn)
 			print(f"Querying: '{query}'")
-			return res
-		except Exception as e:
-			print(e)
-		print('\n')
-
-	def execute(self, query:str) -> list:
-		"""
-		Execute a query on the db
-
-		Args:
-				query (str): query string
-
-		Returns:
-				list: returned from query
-		"""
-		try:
-			res = self.curs.execute(query).fetchall()
-			print(f"'{query}' executed")
 			return res
 		except Exception as e:
 			print(e)
