+++ dash/callbacks.py
@@ -1,19 +1,18 @@
-from app import app
-from dash import Input, Output, html, dcc
+from dash import Input, Output, callback, html, dcc
 from visualisations import create_visualisation
-from index import page_references
 from SQLConnection import SQLConnection
+from datetime import datetime
 
 # SQL connection variable
 # sql = SQLConnection('./ec2-dash/dash.db')
 
-# Call backs for updating the components once a second
-@app.callback(
+# Update current time
+@callback(
 	Output(
-		"rpm_graph",'figure'
+		"current_date",'children'
 	),
 	Output(
-		"rpm_text",'children'
+		"current_time",'children'
 	),
 	[
 		Input(
@@ -21,13 +20,73 @@ from SQLConnection import SQLConnection
 		)
 	]
 )
-def update_rpm_figure(n):
-	# df = sql.query("""
-	# 		SELECT rpm FROM 
-	# 	""")
-	return create_visualisation(df, range(n+1), 'rpm')
+def update_current_time(n):
+	return datetime.now().strftime("%d/%m/%Y"),datetime.now().strftime("%H:%M")
 
-# @app.callback(
+# Update page label button
+@callback(
+	Output(
+		'view_switch', 'children'
+	),
+	Output(
+		'page_link', 'href'
+	),
+	[
+		Input(
+			'view_switch', 'n_clicks'
+		)
+	]
+)
+def change_link(n):
+	print(n)
+	if n % 2 == 0:
+		return 'CURRENT', '/recent'
+	else:
+		return 'RECENT', '/'
+
+# # Update user info at the start of a new ride
+# @callback(
+# 	Output(
+# 		'name', 'children'
+# 	),
+# 	Output(
+# 		'age', 'children',
+# 	),
+# 	Output(
+# 		'gender', 'children',
+# 	),
+# 	[
+# 		Input(
+# 			'interval_component', 'n_intervals'
+# 		)
+# 	]
+# )
+# def update_user_info(n):
+# 	pass
+
+# # # Call backs for updating the components once a second
+# # RPM figure
+# @callback(
+# 	Output(
+# 		"rpm_graph",'figure'
+# 	),
+# 	Output(
+# 		"rpm_text",'children'
+# 	),
+# 	[
+# 		Input(
+# 			'interval_component', 'n_intervals'
+# 		)
+# 	]
+# )
+# def update_rpm_figure(n):
+# 	# df = sql.query("""
+# 	# 		SELECT rpm FROM 
+# 	# 	""")
+# 	return create_visualisation(df, range(n+1), 'rpm')
+
+# # Heart Rate figure
+# @callback(
 # 	Output(
 # 		"heart_rate_graph",'figure'
 # 	),
